+++ src/sasl.rs
@@ -2,7 +2,7 @@ use bytestring::ByteString;
 use either::Either;
 use futures::future::ok;
 use futures::{SinkExt, StreamExt};
-use http::Uri;
+use ntex::http::Uri;
 use ntex::codec::{AsyncRead, AsyncWrite, Framed};
 use ntex::connect::Connect as TcpConnect;
 use ntex::service::{apply_fn, pipeline, IntoService, Service};